+++ toys-core/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * CHANGED: Replaced Toys::Definition::DataFinder with Toys::Definition::SourcePath.
 * IMPROVED: Context directory is now available to tools.
+* IMPROVED: Optionally omit hidden subtools (i.e. names beginning with underscore)
+  from help screens.
 
 ### 0.5.0 / 2018-10-07